:bath: Add cypress itorator WIP
